rag_api/requirements.txt
Flask
selenium
beautifulsoup4
groq
langchain
python-dotenv
langchain-groq
gunicorn
langchain_openai
langchain_postgres
boto3
psycopg2
